moved note into useNoteHandlers
